// types/types.ts
export type Product = {
	id: number
	title: string
	price: number
	description: string
	category: string
	image: string
	rating: {
		rate: number
		count: number
	}
}

export type User = {
	id: number
	email: string
	username: string
	password: string
	name: {
		firstname: string
		lastname: string
	}
	address: {
		city: string
		street: string
		number: number
		zipcode: string
		geolocation: {
			lat: string
			long: string
		}
	}
	phone: string
}